pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spectra"
version = "0.1.0"
description = "Security scanning SaaS backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "arq>=0.26,<1",
    "redis>=5.2,<6",
    "httpx>=0.28,<1",
    "minio>=7.2,<8",
    "anthropic>=0.42,<1",
    "pyjwt[crypto]>=2.10,<3",
    "structlog>=24.4,<25",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-httpx>=0.35,<1",
    "factory-boy>=3.3,<4",
    "aiosqlite>=0.20,<1",
    "ruff>=0.8,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
